added jwt for user

+++ src/users/users.controller.js
@@ -8,11 +8,15 @@ import {
   userInfoSchema
 } from './schemas/schemas.js';
 import { UserService } from './users.service.js';
+import { RoleService } from '../roles/roles.service.js';
+import { HTTPError } from '../errors/http-error.class.js';
+import { verifyJwt, signJwt } from '../common/verify.jwt.js';
 
 export class UserController extends BaseController {
   constructor(logger, prismaService) {
     super(logger);
     this.userService = new UserService(prismaService);
+    this.roleService = new RoleService(prismaService);
     this.bindRoutes([
       {
         path: '/register',
@@ -34,45 +38,75 @@ export class UserController extends BaseController {
         middlewares: [validationMiddleware(deleteSchema)]
       },
       {
-        path: '/update/:id',
+        path: '/update',
         method: 'put',
         func: this.updateUser,
         middlewares: [validationMiddleware(updateSchema)]
-      },
-      {
-        path: '/:id',
-        method: 'get',
-        func: this.getUser,
-        middlewares: [validationMiddleware(userInfoSchema)]
       }
     ]);
   }
 
-  async getUser(req, res, next) {
-    const result = await this.userService.getUserInfo(parseInt(req.params.id, 10));
-    this.ok(res, result);
-  }
-
   async deleteUser(req, res, next) {
+    const authHeader = req.headers.authorization;
+    const token = authHeader && authHeader.split(' ')[1];
+
+    if (!token) return this.send(res, 401, 'No token provided');
+
+    try {
+      req.user = await verifyJwt(token);
+    } catch (err) {
+      return this.send(res, 403, err);
+    }
+
+    const role = await this.roleService.getRoleInfo(req.user.roleId);
+    if (role.name !== 'admin')
+      return this.send(res, 403, 'You have no privelege to delete this user');
+
     const result = await this.userService.deleteUser(parseInt(req.params.id, 10));
-    this.ok(res, result);
+    await this.roleService.deleteRole(parseInt(result.roleId, 10));
+    return this.ok(res, result);
   }
 
   async updateUser(req, res, next) {
-    const result = await this.userService.updateUser(
-      parseInt(req.params.id, 10),
-      req.body
-    );
-    this.ok(res, result);
+    const authHeader = req.headers.authorization;
+    const token = authHeader && authHeader.split(' ')[1];
+
+    if (!token) return this.send(res, 401, 'No token provided');
+
+    try {
+      req.user = await verifyJwt(token);
+    } catch (err) {
+      return this.send(res, 403, err);
+    }
+
+    const result = await this.userService.updateUser(parseInt(req.user.id, 10), req.body);
+    const newToken = await signJwt(result);
+    return this.ok(res, { accessToken: newToken });
   }
 
   async login(req, res, next) {
     const result = await this.userService.loginUser(req.body);
-    this.ok(res, result);
+    if (!result) return this.send(res, 403, 'Email or password is incorrect');
+
+    const token = await signJwt(result);
+    return this.ok(res, { accessToken: token });
   }
 
   async register(req, res, next) {
+    const role = await this.roleService.createRole({ name: 'user' });
+    req.body.roleId = role.id;
     const result = await this.userService.createUser(req.body);
-    this.ok(res, result);
+
+    if (!result)
+      return this.send(res, 409, `User with email: ${req.body.email} already exists`);
+
+    try {
+      const token = await signJwt(result, process.env.ACCESS_TOKEN_SECRET, {
+        expiresIn: '30m'
+      });
+      return this.ok(res, { accessToken: token });
+    } catch (err) {
+      return this.send(res, 403, err);
+    }
   }
 }
